Restructure a package
