--- compare_template_text/__init__.py
@@ -1,1 +1,0 @@
-from configuration.config import PACKAGE_PATH
