+++ examples/compare_template_text.py
