+++ examples/normalize.py
